sample-app/python/requirements.txt
aws-cdk-lib==2.139.0
constructs>=10.3.0,<11.0.0
